ci: add `build.yml` & add `publish.yml`
This commit adds two workflows, build.yml for  building the
crate and publish.yml for releasing the crate to crates.io.

It is important to mention that the build.yml runs rustfmt
as a check alongside the build. However, we may want to
extract this out into an action that automatically formats
instead of merely informing the user that they have not
formatted their code.